Added release-helper script that removes all makefiles before building installer. Various refactorizations.
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@459 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ src/env.cpp
@@ -315,7 +315,7 @@ void CEnvironment::DrawFog () {
 
 	// --------------- calculate the planes ---------------------------
 
-	float slope = tan (ANGLES_TO_RADIANS (Course.GetCourseAngle()));
+	double slope = tan (ANGLES_TO_RADIANS (Course.GetCourseAngle()));
 //	TPlane left_edge_plane = MakePlane (1.0, 0.0, 0.0, 0.0);
 //	TPlane right_edge_plane = MakePlane (-1.0, 0.0, 0.0, Course.width);
