+++ src/states.h
@@ -61,7 +61,7 @@ public:
 	virtual void Keyb(unsigned int key, bool special, bool release, int x, int y) {}
 	virtual void Mouse(int button, int state, int x, int y) {}
 	virtual void Motion(int x, int y) {}
-	virtual void Jaxis(int axis, double value) {}
+	virtual void Jaxis(int axis, float value) {}
 	virtual void Jbutt(int button, int state) {}
 	virtual void Keyb_spec(SDL_keysym sym, bool release) {}
 	virtual void Exit() {}
